requirements.txt
fastapi
uvicorn[standard]
torch
transformers
newsapi
newspaper3k